Update README.md
The theme of this CTF challenge is based on the Rick and Morty TV show. Unfortunately, Rick has been transformed into a pickle and it is up to us to turn him back to his original human form! To accomplish this, we must find 3 ingredients and make a potion 🧪!

The requirements of this CTF is to exploit a web server and find these 3 ingredients (flags). For this writeup, I will be using Kali Linux and some tools that come pre-installed with it. I will begin this writeup after connecting to the target machine. With the background for the CTF out of the way, let’s move on to the writeup.